Make rCRS N removal optional and default to True
Now a parameter to Chromosome.__init__() defaults to True
meaning the N in the rCRS MT sequence is removed from the
sequence returned. This restores the original behavior by
default, but allows a user to request that the N not be
removed by setting RCRS_N_remove=False.

Tests reverted, new test added to test new parameter.